@@ -11,10 +11,12 @@ import PublicSharpIcon from '@mui/icons-material/PublicSharp';
 import SyncOutlinedIcon from '@mui/icons-material/SyncOutlined';
 import ThreeDRotationIcon from '@mui/icons-material/ThreeDRotation';
 import VideoCameraBackIcon from '@mui/icons-material/VideoCameraBackOutlined';
-import WebRtcWindow from '../WebRtcWindow/WebRtcWindow';
-import { WebSocketContext } from '../WebSocket/WebSocket';
-
-const msgpack = require('msgpack-lite');
+import { isEqual } from 'lodash';
+import RenderWindow from '../RenderWindow/RenderWindow';
+import {
+  makeThrottledMessageSender,
+  ViserWebSocketContext,
+} from '../WebSocket/ViserWebSocket';
 
 function CameraToggle() {
   const dispatch = useDispatch();
@@ -100,7 +102,7 @@ function TransformIcons(props) {
   );
 }
 
-// manages a camera and the web rtc stream...
+// manages a camera
 export default function ViewerWindow(props) {
   const sceneTree = props.sceneTree;
   const scene = sceneTree.object;
@@ -108,7 +110,7 @@ export default function ViewerWindow(props) {
   const labelRenderer = sceneTree.metadata.labelRenderer;
 
   const myRef = useRef(null);
-  const websocket = useContext(WebSocketContext).socket;
+  const viser_websocket = useContext(ViserWebSocketContext);
   const field_of_view = useSelector(
     (state) => state.renderingState.field_of_view,
   );
@@ -162,25 +164,24 @@ export default function ViewerWindow(props) {
   };
 
   useEffect(() => {
-    function update_dimensions() {
+    const handleNewDimensions = () => {
       setDimensions({
         height: get_window_height(),
         width: get_window_width(),
       });
-    }
+    };
 
-    window.addEventListener('resize', update_dimensions);
-
-    // force a rerender
     setDimensions({
       height: get_window_height(),
       width: get_window_width(),
     });
     render();
 
+    window.addEventListener('resize', handleNewDimensions);
     return () => {
-      window.removeEventListener('resize', update_dimensions);
+      window.removeEventListener('resize', handleNewDimensions);
     };
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
   // start the three.js rendering loop
@@ -188,6 +189,7 @@ export default function ViewerWindow(props) {
   useEffect(() => {
     myRef.current.append(renderer.domElement);
     render();
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
   const render_height = useSelector(
@@ -238,27 +240,34 @@ export default function ViewerWindow(props) {
     sceneTree.metadata.camera.fov = 50;
   }
 
+  let old_camera_matrix = null;
+  let is_moving = false;
+  const sendThrottledCameraMessage = makeThrottledMessageSender(
+    viser_websocket,
+    25,
+  );
   // update the camera information in the python server
   const sendCamera = () => {
-    if (websocket.readyState === WebSocket.OPEN) {
-      const cmd = 'write';
-      const path = 'renderingState/camera';
-      const data_packet = sceneTree.metadata.camera.toJSON();
-      data_packet.object.timestamp = +new Date();
-      if (camera_choice === 'Render Camera') {
-        data_packet.object.camera_type = camera_type;
+    if (isEqual(old_camera_matrix, sceneTree.metadata.camera.matrix.elements)) {
+      if (is_moving) {
+        is_moving = false;
       } else {
-        data_packet.object.camera_type = 'perspective';
+        return;
       }
-      data_packet.object.render_aspect = render_aspect;
-      const data = {
-        type: cmd,
-        path,
-        data: data_packet,
-      };
-      const message = msgpack.encode(data);
-      websocket.send(message);
+    } else {
+      is_moving = true;
     }
+    old_camera_matrix = sceneTree.metadata.camera.matrix.elements.slice();
+    sendThrottledCameraMessage({
+      type: 'CameraMessage',
+      aspect: sceneTree.metadata.camera.aspect,
+      render_aspect,
+      fov: sceneTree.metadata.camera.fov,
+      matrix: old_camera_matrix,
+      camera_type,
+      is_moving,
+      timestamp: +new Date(),
+    });
   };
 
   // keep sending the camera often
@@ -270,12 +279,31 @@ export default function ViewerWindow(props) {
     return () => {
       clearInterval(refreshIntervalId);
     };
-  }, [websocket, camera_choice, camera_type, render_aspect]);
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [viser_websocket, camera_choice, camera_type, render_aspect]);
+
+  const isWebsocketConnected = useSelector(
+    (state) => state.websocketState.isConnected,
+  );
+  useEffect(() => {
+    if (isWebsocketConnected) {
+      sendThrottledCameraMessage({
+        type: 'CameraMessage',
+        aspect: sceneTree.metadata.camera.aspect,
+        render_aspect,
+        fov: sceneTree.metadata.camera.fov,
+        matrix: sceneTree.metadata.camera.matrix.elements.slice(),
+        camera_type,
+        is_moving: false,
+        timestamp: +new Date(),
+      });
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [isWebsocketConnected]);
 
   return (
     <>
-      {/* the webrtc viewer needs to know the camera pose */}
-      <WebRtcWindow />
+      <RenderWindow />
       <div className="canvas-container-main" ref={myRef}>
         <div className="ViewerWindow-camera-toggle">
           <CameraToggle />
