@@ -1,19 +1,18 @@
 import { useContext } from 'react';
 import { ReactReduxContext } from 'react-redux';
 
-export function subscribe_to_changes(selector_fn, fn) {
-  // selector_fn: returns a value from the redux state
-  // fn_valid: function to run on a valid input
-  // fn_null: function to run on a null input
+export function subscribe_to_changes(selector_fn, callback_fn) {
+  // eslint-disable-next-line react-hooks/rules-of-hooks
   const { store } = useContext(ReactReduxContext);
 
-  let current;
+  let previousState;
   const handleChange = () => {
-    const previous = current;
-    current = selector_fn(store.getState());
-    if (previous !== current) {
-      fn(previous, current);
+    const currentState = selector_fn(store.getState());
+    if (previousState !== currentState) {
+      callback_fn(previousState, currentState);
+      previousState = currentState;
     }
   };
+
   store.subscribe(handleChange);
 }
