@@ -6,9 +6,7 @@ from __future__ import annotations
 
 import time
 from dataclasses import dataclass, field, fields
-from functools import partial
 from pathlib import Path
-from typing import Optional, Literal
 
 import tyro
 from rich.console import Console
@@ -18,8 +16,7 @@ from nerfstudio.engine.trainer import TrainerConfig
 from nerfstudio.pipelines.base_pipeline import Pipeline
 from nerfstudio.utils import writer
 from nerfstudio.utils.eval_utils import eval_setup
-from nerfstudio.utils.writer import EventName, TimeWriter
-from nerfstudio.viewer.server import viewer_utils
+from nerfstudio.viewer.server.viewer_state import ViewerState
 
 CONSOLE = Console(width=120)
 
@@ -29,7 +26,6 @@ class ViewerConfigWithoutNumRays(ViewerConfig):
     """Configuration for viewer instantiation"""
 
     num_rays_per_chunk: tyro.conf.Suppress[int] = -1
-    start_train: tyro.conf.Suppress[bool] = False
 
     def as_viewer_config(self):
         """Converts the instance to ViewerConfig"""
@@ -44,77 +40,59 @@ class RunViewer:
     """Path to config YAML file."""
     viewer: ViewerConfigWithoutNumRays = field(default_factory=ViewerConfigWithoutNumRays)
     """Viewer configuration"""
-    mode: Optional[Literal["bg", "fg"]] = None
-    scene_box_size_ratio: float = 1
 
     def main(self) -> None:
         """Main function."""
-        config, pipeline, _ = eval_setup(
+        config, pipeline, _, step = eval_setup(
             self.load_config,
             eval_num_rays_per_chunk=None,
             test_mode="test",
         )
         num_rays_per_chunk = config.viewer.num_rays_per_chunk
-        mode = self.mode
-        scene_scale = self.scene_box_size_ratio
-        # del self.viewer.mode, self.viewer.scene_box_size_ratio
-
         assert self.viewer.num_rays_per_chunk == -1
         config.vis = "viewer"
         config.viewer = self.viewer.as_viewer_config()
         config.viewer.num_rays_per_chunk = num_rays_per_chunk
 
-        if self.mode is not None:
-            if not hasattr(pipeline._model, "get_outputs_fg_bg"):
-                raise RuntimeError(f"{pipeline._model} does not have foreground background function")
-            print(f"changing rendering to {mode}")
-            pipeline._model.get_outputs = partial(pipeline._model.get_outputs_fg_bg, mode=self.mode,
-                                                  scene_box_size_ratio=self.scene_box_size_ratio)
-
-        self._start_viewer(config, pipeline)
-
-    def _start_viewer(self, config: TrainerConfig, pipeline):
-        base_dir = config.get_base_dir()
-        viewer_log_path = base_dir / config.viewer.relative_log_filename
-        viewer_state, banner_messages = viewer_utils.setup_viewer(
-            config.viewer, log_filename=viewer_log_path, datapath=config.pipeline.datamanager.dataparser.data
-        )
-
-        # We don't need logging, but writer.GLOBAL_BUFFER needs to be populated
-        config.logging.local_writer.enable = False
-        writer.setup_local_writer(config.logging, max_iter=config.max_num_iterations, banner_messages=banner_messages)
-
-        assert viewer_state and pipeline.datamanager.train_dataset
-        viewer_state.init_scene(
-            dataset=pipeline.datamanager.train_dataset,
-            start_train=False,
-        )
-        while True:
-            viewer_state.vis["renderingState/isTraining"].write(False)
-            self._update_viewer_state(viewer_state, config, pipeline)
-
-    def _update_viewer_state(self, viewer_state: viewer_utils.ViewerState, config: TrainerConfig, pipeline: Pipeline):
-        """Updates the viewer state by rendering out scene with current pipeline
-        Returns the time taken to render scene.
-
-        """
-        # NOTE: step must be > 0 otherwise the rendering would not happen
-        step = 1
-        num_rays_per_batch = config.pipeline.datamanager.train_num_rays_per_batch
-        with TimeWriter(writer, EventName.ITER_VIS_TIME) as _:
-            try:
-                viewer_state.update_scene(self, step, pipeline.model, num_rays_per_batch)
-            except RuntimeError:
-                time.sleep(0.03)  # sleep to allow buffer to reset
-                assert viewer_state.vis is not None
-                viewer_state.vis["renderingState/log_errors"].write(
-                    "Error: GPU out of memory. Reduce resolution to prevent viewer from crashing."
-                )
+        _start_viewer(config, pipeline, step)
 
     def save_checkpoint(self, *args, **kwargs):
         """
         Mock method because we pass this instance to viewer_state.update_scene
         """
+
+
+def _start_viewer(config: TrainerConfig, pipeline: Pipeline, step: int):
+    """Starts the viewer
+
+    Args:
+        config: Configuration of pipeline to load
+        pipeline: Pipeline instance of which to load weights
+        step: Step at which the pipeline was saved
+    """
+    base_dir = config.get_base_dir()
+    viewer_log_path = base_dir / config.viewer.relative_log_filename
+    viewer_state = ViewerState(
+        config.viewer,
+        log_filename=viewer_log_path,
+        datapath=pipeline.datamanager.get_datapath(),
+        pipeline=pipeline,
+    )
+    banner_messages = [f"Viewer at: {viewer_state.viewer_url}"]
+
+    # We don't need logging, but writer.GLOBAL_BUFFER needs to be populated
+    config.logging.local_writer.enable = False
+    writer.setup_local_writer(config.logging, max_iter=config.max_num_iterations, banner_messages=banner_messages)
+
+    assert viewer_state and pipeline.datamanager.train_dataset
+    viewer_state.init_scene(
+        dataset=pipeline.datamanager.train_dataset,
+        train_state="completed",
+    )
+    viewer_state.viser_server.set_training_state("completed")
+    viewer_state.update_scene(step=step)
+    while True:
+        time.sleep(0.01)
 
 
 def entrypoint():
